+++ src/Calculator.cpp
@@ -1,31 +1,31 @@
 #include "Calculator.hpp"
 #include "exceptions.hpp"
 
-double Calculator::add(const double &a, const double &b) {
+long double Calculator::add(const long double &a, const long double &b) {
     return _saveLastResult(a + b);
 }
 
-double Calculator::sub(const double &a, const double &b) {
+long double Calculator::sub(const long double &a, const long double &b) {
     return _saveLastResult(a - b);
 }
 
-double Calculator::mul(const double &a, const double &b) {
+long double Calculator::mul(const long double &a, const long double &b) {
     return _saveLastResult(a * b);
 }
 
-double Calculator::div(const double &a, const double &b) {
+long double Calculator::div(const long double &a, const long double &b) {
     if (b == 0) {
         throw DivizionByZeroException("You try to devide by 0");
     }
     return _saveLastResult(a / b);
 }
 
-double Calculator::_saveLastResult(double result) {
+long double Calculator::_saveLastResult(long double result) {
     lastResult = result;
     return result;
 }
 
-double Calculator::getLastResult() const {
+long double Calculator::getLastResult() const {
     if (lastResult.has_value()){
         return *lastResult;
     } else {
